Add handling of the value returned by do_truncate in mdunlink_ao_perFile

+++ src/backend/access/appendonly/aomd.c
@@ -252,7 +252,8 @@ mdunlink_ao_perFile(const int segno, void *ctx)
 
 	sprintf(segPathSuffixPosition, ".%u", segno);
 
-	do_truncate(segPath);
+	if (do_truncate(segPath) < 0 && errno == ENOENT)
+		return false;
 
 	if (unlink(segPath) != 0)
 	{
